Refactored so Flax is now it's own library module, added Hello Flax.
Hello Flax is a sample of how Flax can be used. It allows you to click
on a button which will increment a count on some model. When that model
is updated then the UI is updated to show how many times the button has
been clicked. Config changes work correctly and disposables should be
handled correctly (but maybe not?)

Also refactored Flax to be in it’s own library module. The gradle
config still needs to be cleaned up though.